rebase + refactor
