+++ src/main/kotlin/io/epirus/console/openapi/project/erc777/ERC777Utils.kt
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.epirus.console.token.erc777
+package io.epirus.console.openapi.project.erc777
 
 import io.epirus.console.project.templates.TemplateReader
 import java.io.IOException
@@ -24,16 +24,16 @@ object ERC777Utils {
      */
     fun copy(outputDir: String) {
             val erc777ResourcePath = Paths.get("tokens", "erc777").toString()
-            copyDependency(Paths.get(erc777ResourcePath, "ERC777Token.sol"), Paths.get(outputDir, "ERC777Token.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "Address.sol"), Paths.get(outputDir, "Address.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "Context.sol"), Paths.get(outputDir, "Context.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "ERC777.sol"), Paths.get(outputDir, "ERC777.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "IERC20.sol"), Paths.get(outputDir, "IERC20.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "IERC777.sol"), Paths.get(outputDir, "IERC777.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "IERC777Recipient.sol"), Paths.get(outputDir, "IERC777Recipient.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "IERC777Sender.sol"), Paths.get(outputDir, "IERC777Sender.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "IERC1820Registry.sol"), Paths.get(outputDir, "IERC1820Registry.sol"))
-            copyDependency(Paths.get(erc777ResourcePath, "SafeMath.sol"), Paths.get(outputDir, "SafeMath.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "ERC777Token.sol"), Paths.get(outputDir, "ERC777Token.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "Address.sol"), Paths.get(outputDir, "Address.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "Context.sol"), Paths.get(outputDir, "Context.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "ERC777.sol"), Paths.get(outputDir, "ERC777.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "IERC20.sol"), Paths.get(outputDir, "IERC20.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "IERC777.sol"), Paths.get(outputDir, "IERC777.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "IERC777Recipient.sol"), Paths.get(outputDir, "IERC777Recipient.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "IERC777Sender.sol"), Paths.get(outputDir, "IERC777Sender.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "IERC1820Registry.sol"), Paths.get(outputDir, "IERC1820Registry.sol"))
+        copyDependency(Paths.get(erc777ResourcePath, "SafeMath.sol"), Paths.get(outputDir, "SafeMath.sol"))
     }
 
     @Throws(IOException::class)
