+++ src/main/kotlin/io/epirus/console/openapi/subcommands/JarOpenApiCommand.kt
@@ -91,13 +91,11 @@ class JarOpenApiCommand : AbstractOpenApiCommand() {
             Paths.get(projectDirectoryPath.toString(), projectOptions.projectName).toString(),
             withOpenApi = true,
             withSwaggerUi = true,
-            withShadowJar = true,
-            withClientJar = true,
-            withServerJar = true)
+            withShadowJar = true)
 
         Files.copy(
             getJarFile(projectDirectoryPath),
-            File(projectOptions.outputDir, "${projectOptions.projectName}$JARSUFFIX").toPath(),
+            File(projectOptions.outputDir, "${projectOptions.projectName}$JAR_SUFFIX").toPath(),
             StandardCopyOption.REPLACE_EXISTING
         )
